# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kioserra API",
        "description": "API documentation for Kioserra backend",
        "contact": {
            "email": "support@kioserra.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/subscriptions": {
            "get": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Get list of subscriptions",
                "operationId": "b6cad0cd2df2d41298b989a858d779db",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Create a new subscription",
                "operationId": "4b05f22cd1772563f2c2ef53e04176ea",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Subscription"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Subscription created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
        "/subscriptions/{id}": {
            "get": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Get a subscription by ID",
                "operationId": "2bf90b2ad6b7c7cae24661b9b24c855e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Subscription not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Update a subscription",
                "operationId": "705978984b3a523c12a09207b7523801",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Subscription"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subscription updated successfully"
                    },
                    "404": {
                        "description": "Subscription not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Delete a subscription",
                "operationId": "a66a426f38da641ada92b9cb7afde5ce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Subscription deleted successfully"
                    },
                    "404": {
                        "description": "Subscription not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Subscription": {
                "required": [
                    "store_id",
                    "start_date",
                    "end_date",
                    "amount_paid",
                    "payment_method"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "store_id": {
                        "type": "integer"
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "amount_paid": {
                        "type": "number",
                        "format": "float"
                    },
                    "payment_method": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}